Add Updater trait that is implemented by WalletUpdater
