+++ rust/src/wallet/mod.rs
@@ -1,5 +1,7 @@
 pub mod recorded;
 pub mod utils;
 mod wallet;
+mod updater;
 
 pub use wallet::SpWallet;
+pub use updater::WalletUpdater;
